requirements.txt
Django==5.1.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0

